requirements.txt
streamlit
openai
edge-tts
moviepy
pillow
python-dotenv
requests
